Removing half-pixel correction fixed font artifacts, weird.

+++ src/font.h
@@ -37,7 +37,7 @@ struct Font {
 };
 typedef struct Font Font;
 
-void font_init(Font *font, FT_Library *ft, const char *filename, int point_size, int hdpi, int vdpi);
+void font_init(Font *font, FT_Library *ft, const char *filename, int point_size, float hdpi, float vdpi);
 void font_destroy(Font *font);
 
 int font_contains_char(Font *font, char c);
